libpam: simplify IF_NO_PAMH
The first argument of IF_NO_PAMH is supposed to be the name of the
function which was called with pamh being NULL.

With __FUNCTION__ the name can be inserted automatically by the
compiler which is also already done with D macro.

Fixes a bug in which _pam_drop_env erroneously logs with the
function name _pam_make_env.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>