pam_echo: handle short reads
If the file parsed by pam_echo is larger than INT_MAX, then it is
possible that uninitialized memory is printed on screen.

The return value of pam_modutil_read is not negative if the size
argument (casted to an int) is negative. Instead 0 is returned.

This can also happen with any other file if a short read is triggered.

Check if file is fully parsed or not.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>