pam_unix: handle invalid names in _unix_getpwnam
It is possible to trigger an out of boundary read with very long
usernames (strlen's result is stored in an int) or, with even
longer usernames, match other users with same prefix. This would
mean that roott[and lots of t's following] could match root user.

Also do not allow ':' in names when iterating through the passwd
file this way.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>