pam_unix: fix possible shadow signed overflows
It is possible to trigger signed integer overflows in
check_shadow_expiry if /etc/shadow contains very large values.

Since these values have to be set by a system administrator, it would
already count as a configuration error.

Yet, avoid overflows which would consider accounts which are supposed
to be valid for a veeery long time as already invalid. Also, it would
be undefined behavior for almost all C standards.

Also consider every negative value as invalid, not just -1. The shadow
project has different ways of handling these values, but this approach
is in sync with its lib/isexpired.c implementation.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>