pam_limits: avoid sscanf in parse_config_file
Even though sscanf is easy to use for scanning strings, it has the
drawback that the required memory has to be allocated beforehand.

Since variable "line" is not accessed anymore after sscanf, it is
easier to point into the memory assigned to "line". The function
strtok_r can be used as a replacement for sscanf.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>