pam_namespace: validate amount of uids in config
If more than INT_MAX uids are found in a configuration line, the
variable `count` would trigger a signed integer overflow.

If more than UINT_MAX uids are found in a configuration line, then
the `num_uids` counter is invalid, which could eventually lead to out
of boundary accesses.

Also make sure that size multiplication for malloc does not overflow.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>