pam_unix: do not truncate user names
This could allow users with very long names to impersonate a user
with a 255 characters long name.

The check if the argument argv[1] actually matches the user name
implies that "user" can unconditionally be set to argv[1]: If they are
equal, the strings are obviously equal. If they are not or if null is
returned by getuidname, "user" is set to argv[1] anyway.

This way, the static buffer can be safely removed because the result
of getpwuid() is not stored, which means that subsequent calls to
such functions can safely overwrite their internal buffers.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>