configure: fail if specified option cannot be satisfied
The options whether to build with support for libprelude, libaudit,
libselinux, or libeconf are set to enable-if-available.  These options
also have a configure flag `--enable-foo`, which currently fall back to
the feature being disabled if not available.

Change these feature flags to fail if specified explicitly and the
required dependencies cannot be satisfied.

Prompted by #728 and #746