Unify PAM return value usage
PAM defines a set of common return-values, like PAM_SUCCESS or
PAM_AUTH_ERR, which are used internally and in public API.  Commonly
these return-values are stored in variables named `retval`.

Unify the naming of variables such that if-and-only-if a variable is
named `retval` it holds a PAM style return-value.

Also use explicitly PAM_SUCESS instead of the value 0 in assignments and
comparisons.