libpam: introduce pam_alloc
The functions pam_add_alloc and pam_free_alloc manage unnamed data which
has to be freed when pam_end is called.

Use pam_alloc for data which is never looked up again but has to remain
in heap for later uses. This can be a good foundation for data
management of reentrant function implementations.

Signed-off-by: Tobias Stoeckmann <tobias@stoeckmann.org>