doc: fix pam_get_data manual regarding PAM_NO_MODULE_DATA return value
pam_get_data manual page used to claim that PAM_NO_MODULE_DATA is
returned not only when no module specific data is present but also
when there is a corresponding data entry with NULL value.

As this claim is in contradiction with the implementation and with the
pam manual page, fix pam_get_data manual page.

* doc/man/pam_get_data.3.xml (PAM_NO_MODULE_DATA): Sync with
the implementation and with the pam manual page.

Resolves: https://github.com/linux-pam/linux-pam/pull/780