Set default fail delay to 0.5 seconds
The default fail delay is 2 seconds (+/- 50% randomly). This is done to
prevent brute force attacks, but unfortunately because isn't applied
exponentially, it also greatly degrades the UX for legitimate users.
Every typo is immediately met with an annoying 2 second delay.

A full fix would be to persist the login failure times and apply an
exponential delay (this is what Windows does), however this is quite
complicated due to PAM's architecture (see the `faillock` plugin for
an example of this).

This simple fix greatly improves the UX at very minimal security cost.
Any brute force attack that can tolerate a 0.5 second delay will also
work with a 2 second delay. In other words a factor of 4 is nothing
from an attacker's perspective, but makes a big difference from a
legitimate user's perspective.

Fixes #778