modules/pam_succeed_if: Add support for evaluating pam env variables
It may be useful in services files to be able to skip modules or take
different paths depending on variables that may have been set by a
previous module in the stack.

So make it possible for pam_succeed_if to get an env variable and do
comparisons depending on its value.

Since variables can also be just set or unset, introduce a new syntax to
perform such check, such as:

  env:VARIABLE is unset
  env:VARIABLE is set