pam_inline: introduce pam_asprintf(), pam_snprintf(), and pam_sprintf()
pam_asprintf() is essentially asprintf() with the following semantic
difference: it returns the string itself instead of its length.

pam_snprintf() is essentially snprintf() with the following semantic
difference: it returns -1 in case of truncation.

pam_sprintf() is essentially snprintf() but with a check that the buffer
is an array, and with an automatically calculated buffer size.

Use of these helpers would make error checking simpler.