pam_namespace: fix potential privilege escalation
Existing protection provided by protect_dir() and protect_mount() were
bind mounting on themselves all directories part of the to-be-secured
paths. However, this works *only* against attacks executed by processes
in the same mount namespace as the one the mountpoint was created in.
Therefore, a user with an out-of-mount-namespace access, or multiple
users colluding, could exploit multiple race conditions, and, for
instance, elevate their privileges to root.

This commit keeps the existing protection as a defense in depth
measure, and to keep the existing behavior of the module. However,
it converts all the needed function calls to operate on file
descriptors instead of absolute paths to protect against race
conditions globally.

Signed-off-by: Olivier Bal-Petre <olivier.bal-petre@ssi.gouv.fr>
Signed-off-by: Dmitry V. Levin <ldv@strace.io>