pam_nologin: don't display message to root
Showing the "come back later" message to root is confusing.  Let's stop
doing that.

Showing root the message (at the INFO level rather than ERR) seems to be
an intended feature that has been around since at least the "Initial commit" in
this repo in 2000, in this form:

```
       if (user_pwd && user_pwd->pw_uid == 0) {
         message.msg_style = PAM_TEXT_INFO;
```

But it's not clear to me if this was ever a good idea or why we would
want it.  It certainly is at odds with the message that systemd writes
to `/run/nologin` telling the user to "come back later", since root is
indeed permitted to login successfully.

Adjust the tests: now that we no longer attempt to read the file when
logging in as root, the "failed to read file (because it's a directory)"
testcase needs to be moved from the uid-0 block to the non-uid-0 block.

Fixes #941