unix_chkpwd: use setregid and setreuid
setgid and setuid only reset the saved set-(group|user)-ID
if the process has CAP_SETGID/CAP_SETUID. Generally this is only the
case when running as root.

If we use setregid/setreuid and pass the current real ID as both the real
and effective ID, the saved set-(group|user)-ID will always be reset,
regardless of the process capabilities.

For example, if the program is set-group-ID 'shadow', it is currently
possible for the process to set its effective group id back to 'shadow'
after calling setgid(rgid). By using setregid(rgid, rgid), we prevent
that from happening.

Also update the "no-return" check to use the previous effective IDs
instead of assuming 0 (root).

Signed-off-by: Mike Gilbert <floppym@gentoo.org>